+++ src/UncertainData.jl
@@ -11,7 +11,7 @@ using KernelDensity
 
 
 display_update = true
-version = "v0.6.0"
+version = "v0.8.0"
 update_name = "update_$version"
 
 if display_update
@@ -23,30 +23,8 @@ printstyled(stdout,
 New features
 ----------------
 
-
-### New features
-
-- `UncertainValue` constructor for vectors of zero-dimensional arrays.
-
-New resampling types for uncertain index-value datasets.
-- `ConstrainedValueResampling` resampling type. Makes it easier to provide sampling 
-    constraints to uncertain value collections, especially if there are many of 
-    them and they need different sampling constraints.
-- `ConstrainedIndexValueResampling` resampling type. Makes it easier to provide 
-    sampling constraints to index-value datasets, especially if there are many of
-    them and they need different sampling constraints.
-- `SequentialResampling` resampling type.
-- `SequentialInterpolationResampling` resampling type. Combines sequential 
-    resampling and interpolation (in that order).
-
-and their resampling methods: 
-
-- `resample(x::AbstractUncertainIndexValueDataset, resampling::ConstrainedValueResampling`.
-- `resample(x::AbstractUncertainIndexValueDataset, resampling::ConstrainedIndexValueResampling`.
-- `resample(x::AbstractUncertainIndexValueDataset, resampling::SequentialResampling`, a
-    method that resamples sequentially according to the indices of `x`.
-- `resample(x::AbstractUncertainIndexValueDataset, resampling::SequentialInterpolatedResampling`, a 
-    method that combines sequential sampling and interpolation (in that order).
+- Added binned resampling methods that uses `BinnedResampling` and 
+    `BinnedMeanResampling` schemes.
 """; color = :light_magenta)
 touch(joinpath(@__DIR__, update_name))
 end