+++ test/resampling/test_resampling_inplace.jl
@@ -65,3 +65,67 @@ resample!(idxs, vals, U)
 
 @test any(isnan.(idxs)) == false
 @test any(isnan.(vals)) == false
+
+
+
+###################################################################################
+# Draw N realisations of an uncertain value into vector-like containers of length N
+###################################################################################
+
+# A single uncertain value resampled multiple times into a N-element vector
+N = 10
+x = fill(NaN, 10)
+
+resample!(x, UncertainValue(Normal(0, 1)))
+@test any(isnan.(x)) == false
+
+# A single uncertain value resampled multiple times into a N-element MVector
+N = 10
+x = MVector{N, Float64}(repeat([NaN], N))
+
+resample!(x, UncertainValue(Normal(0, 1)))
+@test any(isnan.(x)) == false
+
+# A single uncertain value resampled multiple times into a 3-element FieldVector-type
+mutable struct Vector3DType <: FieldVector{3, Float64}
+    x::Float64
+    y::Float64
+    z::Float64
+end
+
+x = Vector3DType(NaN, NaN, NaN)
+
+resample!(x, UncertainValue(Normal(0, 1)))
+@test any(isnan.(x)) == false
+
+#########################################################################################
+# Draw single realisations of N uncertain values into vector-like containers of length N
+#########################################################################################
+
+# Three uncertain values resampled element-wise into a 3-element vector
+x = repeat([NaN], 3)
+resample!(x, (val, val, val))
+@test any(isnan.(x)) == false
+
+# Three uncertain values resampled element-wise into a 3-element MVector
+N = 3
+x = MVector{N, Float64}(repeat([NaN], N))
+uvals = [UncertainValue([0], [1]), UncertainValue([1], [1]), UncertainValue([2], [1])]
+resample!(x, uvals)
+@test any(isnan.(x)) == false
+
+# when the number of elements does not match the number of uncertain values
+x = MVector{2, Float64}(repeat([NaN], 2))
+uval = UncertainValue(Normal(0, 1))
+resample!(x, uval)
+
+# Two uncertain values resampled elementwise into a 2-element vector-like type
+mutable struct VectorLikeType <: FieldVector{2, Float64}
+    x::Float64
+    y::Float64
+end
+
+x = VectorLikeType(NaN, NaN)
+uvals = [UncertainValue([0], [1]), UncertainValue([1], [1])]
+resample!(x, uvals)
+@test any(isnan.(x)) == false