+++ docs/src/changelog.md
@@ -1,6 +1,23 @@
 
 # Changelog
 
+## Uncertaindata.jl v0.8.0
+
+### New functionality
+
+- Added binned resampling methods that uses [`BinnedResampling`](@ref) and [`BinnedMeanResampling`](@ref) schemes.
+    - [`resample(::AbstractUncertainIndexValueDataset, ::BinnedResampling`](@ref)
+    - [`resample(::AbstractUncertainIndexValueDataset, ::BinnedMeanResampling`](@ref)
+
+### Bug fixes
+
+- Fixed bug where `resample!` method for vectors and tuples of uncertain values didn't return the 
+    expected result.
+
+### Improvements
+
+- Improved `resample!` docs.
+
 ## Uncertaindata.jl v0.7.0
 
 ### New functionality